댓글 저장 및 수정, 유저 정보 조회 및 수정 구현
세션 검증 로직 자세히 할 필요 있음

+++ src/main/java/com/baki/backer/domain/auth/AuthController.java
@@ -64,7 +64,7 @@ public class AuthController {
         return ResponseEntity.ok("로그아웃 성공");
     }
     @GetMapping("/admin")
-    public ResponseEntity<?> adminPage(@SessionAttribute(name = "username", required = false) Integer Id){
+    public ResponseEntity<?> adminPage(@SessionAttribute(name = "username", required = false) Long Id){
         if(Id == null){
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인이 필요합니다.");
         }
