+++ src/main/java/com/baki/backer/domain/member/MemberRepository.java
@@ -7,14 +7,14 @@ import org.springframework.lang.NonNull;
 
 import java.util.Optional;
 
-public interface MemberRepository extends JpaRepository<Member,Integer> {
+public interface MemberRepository extends JpaRepository<Member,Long> {
     boolean existsByUsername(String username);
     boolean existsByName(String name);
-    boolean existsById(@NonNull Integer id);
+    boolean existsById(@NonNull Long id);
     Optional<Member> findByUsername(String username);
 
     @Query("select m.id from Member m where m.username = :username")
-    Integer findIdByUsername(@Param("username") String username);
+    Long findIdByUsername(@Param("username") String username);
 
 
 }
