+++ src/main/java/com/baki/backer/domain/post/repository/PostRepository.java
@@ -7,8 +7,8 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post,Integer>,CustomPostRepository {
+public interface PostRepository extends JpaRepository<Post,Long>,CustomPostRepository {
     @Query("SELECT p FROM Post p LEFT JOIN  fetch p.comments where p.post_id = :post_id")
-    Post findPostWithComment(@Param("post_id") Integer post_id);
+    Post findPostWithComment(@Param("post_id") Long post_id);
 }
